requirements.txt
openai-whisper>=20231117
google-genai>=1.0.0
soundcard>=0.4.3
sounddevice>=0.4.0
soundfile>=0.12.1
watchdog>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydub>=0.25.1
